retain session through view changes
